pyproject.toml
[project]
name = "srajob"
version = "0.1.0"
description = "Job board with scraping on scheduled intervals using Temporal.io"
readme = "README.md"
authors = [{ name = "matchaboar", email = "matchaboar@email.com" }]
requires-python = "==3.13.*"
dependencies = [
    "convex>=0.7.0",
    "fetchfox-sdk>=1.0.1",
    "firecrawl>=4.8.0",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "temporalio==1.17.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "ruff>=0.1.0"]
ai = []

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[build-system]
requires = ["uv_build>=0.8.8,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "green>=4.0.2",
    "pyright>=1.1.407",
]
